added per-year indexing for 2008
This is just an attempt to be able to browse through all posts in one
year instead of having to go through them month by month.
For now, the work was only made for 2008, but if it works well future
commits should do the same for the rest of the years (2009-2015).